stat() related changes
git-svn-id: svn://svn.caucho.com/resin/trunk@1241 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/vfs/PathWrapper.java
@@ -628,7 +628,7 @@ public abstract class PathWrapper extends Path {
     getWrappedPath().setObject(obj);
   }
 
-  public int getInode()
+  public long getInode()
   {
     return getWrappedPath().getInode();
   }
